Fix macros template naming
The k8s 1.5 does not support the _ in names of keys
of ConfigMaps, so the 'oslo_messaging' is invalid.

But in the same moment we cant use the "-" as
separator because the file name is used due
to importing in jinja2 like:

  import 'oslo-messaging.j2' as oslo-messaging

the "oslo-messaging" is invalid jinja variable.

So, currently the only way is didn't use any
separator.

Change-Id: I5594db296d745fe57b13ea3d2cd38be1e74d4d10