[WIP] Render vm_memory_high_watermark for RabbitMQ
RabbitMQ/Erlang do not properly determine available RAM when running in
containerised environment, plus default value for watermark is 0.4,
which is low for nodes running RabbitMQ only.

To fix this, we termine correct value of the limit using cgroups:
/sys/fs/cgroup/memory/memory.limit_in_bytes

then determine useful memory based on the ratio (0.75 is a recommended
default) and then render additional variable in RabbitMQ config file.

Some additional information:
https://www.rabbitmq.com/memory.html

Change-Id: I732338ea70c9277961f5f8514a79a247147f543a